refactor: TuripPlace 관련 로직을 TuripRepository로 통합
