+++ android/app/src/main/java/com/on/turip/data/turip/datasource/DefaultTuripRemoteDataSource.kt
@@ -1,6 +1,8 @@
 package com.on.turip.data.turip.datasource
 
 import com.on.turip.core.result.TuripResult
+import com.on.turip.data.place.dto.TuripPlaceOrderRequest
+import com.on.turip.data.place.dto.TuripPlacesResponse
 import com.on.turip.data.result.safeApiCall
 import com.on.turip.data.turip.dto.TuripCreationResponse
 import com.on.turip.data.turip.dto.TuripPatchRequest
@@ -43,5 +45,29 @@ class DefaultTuripRemoteDataSource @Inject constructor(
     override suspend fun getTuripsByPlaceId(placeId: Long): TuripResult<TuripsByPlaceResponse> =
         withContext(coroutineContext) {
             safeApiCall { turipService.getTuripsByPlaceId(placeId) }
+        }
+
+    override suspend fun getTuripPlaces(turipId: Long): TuripResult<TuripPlacesResponse> =
+        safeApiCall { turipService.getTuripPlaces(turipId) }
+
+    override suspend fun createTuripPlace(
+        turipId: Long,
+        placeId: Long,
+    ): TuripResult<Unit> = safeApiCall { turipService.postTuripPlace(turipId, placeId) }
+
+    override suspend fun deleteTuripPlace(
+        turipId: Long,
+        placeId: Long,
+    ): TuripResult<Unit> = safeApiCall { turipService.deleteTuripPlace(turipId, placeId) }
+
+    override suspend fun patchTuripPlacesOrder(
+        turipId: Long,
+        turipPlaceOrderRequest: TuripPlaceOrderRequest,
+    ): TuripResult<Unit> =
+        safeApiCall {
+            turipService.patchTuripPlaceOrder(
+                turipId,
+                turipPlaceOrderRequest,
+            )
         }
 }
