+++ android/app/src/main/java/com/on/turip/data/turip/repository/DefaultTuripRepository.kt
@@ -2,10 +2,13 @@ package com.on.turip.data.turip.repository
 
 import com.on.turip.core.result.TuripResult
 import com.on.turip.core.result.mapCatching
+import com.on.turip.data.place.dto.TuripPlaceOrderRequest
+import com.on.turip.data.place.toDomain
 import com.on.turip.data.turip.datasource.TuripRemoteDataSource
 import com.on.turip.data.turip.toDomain
 import com.on.turip.data.turip.toPatchRequestDto
 import com.on.turip.data.turip.toPostRequestDto
+import com.on.turip.domain.favorite.TuripPlace
 import com.on.turip.domain.folder.Turip
 import com.on.turip.domain.folder.repository.TuripRepository
 import javax.inject.Inject
@@ -31,4 +34,27 @@ class DefaultTuripRepository @Inject constructor(
         turipRemoteDataSource
             .getTuripsByPlaceId(placeId)
             .mapCatching { it.toDomain() }
+
+    override suspend fun loadTuripPlaces(turipId: Long): TuripResult<List<TuripPlace>> =
+        turipRemoteDataSource.getTuripPlaces(turipId).mapCatching { it.toDomain() }
+
+    override suspend fun createTuripPlace(
+        turipId: Long,
+        placeId: Long,
+    ): TuripResult<Unit> = turipRemoteDataSource.createTuripPlace(turipId = turipId, placeId = placeId)
+
+    override suspend fun deleteTuripPlace(
+        turipId: Long,
+        placeId: Long,
+    ): TuripResult<Unit> = turipRemoteDataSource.deleteTuripPlace(turipId = turipId, placeId = placeId)
+
+    override suspend fun updateTuripPlacesOrder(
+        turipId: Long,
+        updatedOrder: List<Long>,
+    ): TuripResult<Unit> =
+        turipRemoteDataSource
+            .patchTuripPlacesOrder(
+                turipId = turipId,
+                turipPlaceOrderRequest = TuripPlaceOrderRequest(turipPlaceIdsOrder = updatedOrder),
+            )
 }
