+++ android/app/src/main/java/com/on/turip/di/DataSourceModule.kt
@@ -8,8 +8,6 @@ import com.on.turip.data.login.datasource.AuthDataSource
 import com.on.turip.data.login.datasource.AuthRemoteDataSource
 import com.on.turip.data.login.datasource.MemberDataSource
 import com.on.turip.data.login.datasource.MemberRemoteDataSource
-import com.on.turip.data.place.datasource.DefaultTuripPlaceRemoteDataSource
-import com.on.turip.data.place.datasource.TuripPlaceRemoteDataSource
 import com.on.turip.data.region.datasource.DefaultRegionRemoteDataSource
 import com.on.turip.data.region.datasource.RegionRemoteDataSource
 import com.on.turip.data.searchhistory.datasource.DefaultSearchHistoryDataSource
@@ -52,12 +50,6 @@ abstract class DataSourceModule {
     @Binds
     @Singleton
     abstract fun bindTuripRemoteDataSource(defaultTuripRemoteDataSource: DefaultTuripRemoteDataSource): TuripRemoteDataSource
-
-    @Binds
-    @Singleton
-    abstract fun bindTuripPlaceRemoteDataSource(
-        defaultTuripPlaceRemoteDataSource: DefaultTuripPlaceRemoteDataSource,
-    ): TuripPlaceRemoteDataSource
 
     @Binds
     @Singleton
