+++ android/app/src/main/java/com/on/turip/domain/folder/repository/TuripRepository.kt
@@ -1,6 +1,7 @@
 package com.on.turip.domain.folder.repository
 
 import com.on.turip.core.result.TuripResult
+import com.on.turip.domain.favorite.TuripPlace
 import com.on.turip.domain.folder.Turip
 
 interface TuripRepository {
@@ -16,4 +17,21 @@ interface TuripRepository {
     suspend fun deleteTurip(turipId: Long): TuripResult<Unit>
 
     suspend fun loadTuripsByPlaceId(placeId: Long): TuripResult<List<Turip>>
+
+    suspend fun loadTuripPlaces(turipId: Long): TuripResult<List<TuripPlace>>
+
+    suspend fun createTuripPlace(
+        turipId: Long,
+        placeId: Long,
+    ): TuripResult<Unit>
+
+    suspend fun deleteTuripPlace(
+        turipId: Long,
+        placeId: Long,
+    ): TuripResult<Unit>
+
+    suspend fun updateTuripPlacesOrder(
+        turipId: Long,
+        updatedOrder: List<Long>,
+    ): TuripResult<Unit>
 }
